Sell TWAP strategy (closes #454) (#455)
* 1 - create sell_twap strategy by copying sell strategy files

* 2 - create sample_selltwap.cfg by copying sample_sell.cfg file

* 3 - added the new config params to sellTwapStrategy.go

* 4 - add makeDowFilter method

* 5 - dowFilter should be an array

* 6 - update sample_selltwap.cfg

* 7 - get filter in GetLevels along with daily cap amount in base units

* 8 - calculate bucketIdx

* 9 - extract to makeBucketInfo() to keep the code clean

* 10 - fill in remaining fields of bucketInfo and roundInfo

* 11 - use time as a seed for randomness

* 12 - use numHoursToSell config value in calculations via totalBucketsTargeted

* 13 - bucketProgress added to bucket String()

* 14 - save bucket and round for future rounds

* 15 - fix: when baseRemaining is less than minOrderSize then return baseRemaining

* 16 - save previousRoundID only instead of previousRound

* 17 - sizeCapped -> sizeBaseCapped

* 18 - create levels from the computed round

* 19 - simplify bucket creation in each round by making cases explicit

* 20 - add fields to dynamicBucketValues: isNew and roundID

* 21 - updated ceilDate() + test to give you same day value (inclusive)

* 22 - fix startTime and endTime on buckets, extract out makeFirstBucketFrame call from cutoverToNewBucket

* 23 - handle date boundaries; include totalBaseSurplusStart field

* 24 - add bucketTimeElapsed to bucketInfo

* 25 - BucketInfo logs date, dayID, day as string, and sizeSeconds

* 26 - totalBucketsTargeted -> totalBucketsToSell

* 29 - add UUID() to the bucket and log for easy tracking

* 30 - add bucketUUID to roundInfo for logging

* 31 - fix secondsElapsedToday in roundInfo

* 32 - refactor out volFilter from the three inner factory methods for bucketInfo

* 33 - TestMakeFirstBucketFrame

* 34 - TestMakeFirstBucketFrame2

* 35 - TestUpdateExistingBucket

* 36 - use a startDate and endDate in tests that is not the floor/ceil values

* 37 - TestCutoverToNewBucketSameDay_InsideNumHoursToSell + update surplus distribution logic

* 38 - TestCutoverToNewBucketSameDay_OutsideNumHoursToSell + fix surplus and base capacity

* 39 - TestCutoverToNewBucketSameDay_WithNewVolumeSold

* 40 - TestFirstDistributionOfBaseSurplus

* 41 - TestBucketInfoString

* 42 - TestBucketInfoUUID

* 43 - return empty levels when sizeBaseCapped == 0.0

* 44 - combine bucketInfo and roundInfo log lines into a single line for better searchability

* 45 - return no levels if sizeBaseCapped <= 0.0

* 46 - fix many numeric bugs by removing cutoverToNewBucketSameDay(), a new bucketInfo frame can have a surplus; all tests updated, converted to table based tests

* 47 - enable filters to run on 'sell_twap' strategy

* 48 - update previous bucket when we cutover and add new isLast param

* 49 - adjust sample_selltwap.cfg to have some more reasonable values

* 50 - add walkthrough for sell_twap strategy and add to README.md

+++ cmd/trade.go
@@ -322,6 +322,7 @@ func makeStrategy(
 	assetQuote hProtocol.Asset,
 	ieif *plugins.IEIF,
 	tradingPair *model.TradingPair,
+	filterFactory *plugins.FilterFactory,
 	options inputs,
 	threadTracker *multithreading.ThreadTracker,
 ) api.Strategy {
@@ -346,6 +347,7 @@ func makeStrategy(
 		*options.stratConfigPath,
 		*options.simMode,
 		botConfig.IsTradingSdex(),
+		filterFactory,
 	)
 	if e != nil {
 		l.Info("")
@@ -364,9 +366,8 @@ func makeBot(
 	exchangeShim api.ExchangeShim,
 	ieif *plugins.IEIF,
 	tradingPair *model.TradingPair,
-	db *sql.DB,
+	filterFactory *plugins.FilterFactory,
 	strategy api.Strategy,
-	assetDisplayFn model.AssetDisplayFn,
 	threadTracker *multithreading.ThreadTracker,
 	options inputs,
 ) *trader.Trader {
@@ -417,19 +418,11 @@ func makeBot(
 			plugins.MakeFilterMakerMode(exchangeShim, sdex, tradingPair),
 		)
 	}
-	if len(botConfig.Filters) > 0 && *options.strategy != "sell" && *options.strategy != "delete" {
+	if len(botConfig.Filters) > 0 && *options.strategy != "sell" && *options.strategy != "sell_twap" && *options.strategy != "delete" {
 		log.Println()
 		utils.PrintErrorHintf("FILTERS currently only supported on 'sell' and 'delete' strategies, remove FILTERS from the trader config file")
 		// we want to delete all the offers and exit here since there is something wrong with our setup
 		deleteAllOffersAndExit(l, botConfig, client, sdex, exchangeShim, threadTracker)
-	}
-	filterFactory := plugins.FilterFactory{
-		ExchangeName:   botConfig.TradingExchangeName(),
-		TradingPair:    tradingPair,
-		AssetDisplayFn: assetDisplayFn,
-		BaseAsset:      assetBase,
-		QuoteAsset:     assetQuote,
-		DB:             db,
 	}
 	for _, filterString := range botConfig.Filters {
 		filter, e := filterFactory.MakeFilter(filterString)
@@ -571,6 +564,14 @@ func runTradeCmd(options inputs) {
 		tradingPair,
 		sdexAssetMap,
 	)
+	filterFactory := &plugins.FilterFactory{
+		ExchangeName:   botConfig.TradingExchangeName(),
+		TradingPair:    tradingPair,
+		AssetDisplayFn: assetDisplayFn,
+		BaseAsset:      assetBase,
+		QuoteAsset:     assetQuote,
+		DB:             db,
+	}
 	strategy := makeStrategy(
 		l,
 		network,
@@ -582,6 +583,7 @@ func runTradeCmd(options inputs) {
 		assetQuote,
 		ieif,
 		tradingPair,
+		filterFactory,
 		options,
 		threadTracker,
 	)
@@ -593,9 +595,8 @@ func runTradeCmd(options inputs) {
 		exchangeShim,
 		ieif,
 		tradingPair,
-		db,
+		filterFactory,
 		strategy,
-		assetDisplayFn,
 		threadTracker,
 		options,
 	)
